chore: add -W flags to compiler

+++ components/progress_bar/progress_bar.cpp
@@ -5,7 +5,7 @@
 
 void ProgressBar::render_component(std::array<Color, MATRIX_SIZE>& matrix) {
     Rectangle border(pos, width, height, primary, Color(0, 0, 0, 0));
-    Rectangle infill(pos, round(progress * (width - 2) / 100) + 2, height, Color(0, 0, 0, 0), secondary);
+    Rectangle infill(pos, static_cast<int>(std::round(progress * (width - 2) / 100.0) + 2), height, Color(0, 0, 0, 0), secondary);
 
     border.render_component(matrix);
     infill.render_component(matrix);
